# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "A股投资平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/accounts/{account_id}/evaluations": {
      "post": {
        "tags": [
          "decisions"
        ],
        "summary": "Evaluate Position",
        "operationId": "evaluate_position_api_v1_accounts__account_id__evaluations_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionEvaluationInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_JobView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}/decisions": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "List Current Decisions",
        "operationId": "list_current_decisions_api_v1_accounts__account_id__decisions_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PositionDecisionPage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decisions/{decision_id}": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "Get Decision",
        "operationId": "get_decision_api_v1_decisions__decision_id__get",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PositionDecision_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decisions/{decision_id}/plans": {
      "post": {
        "tags": [
          "decisions"
        ],
        "summary": "Create Decision Plan",
        "operationId": "create_decision_plan_api_v1_decisions__decision_id__plans_post",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionPlanInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PlanView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}/monitors": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "List Monitors",
        "operationId": "list_monitors_api_v1_accounts__account_id__monitors_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MonitorPage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decisions/{decision_id}/monitor": {
      "put": {
        "tags": [
          "decisions"
        ],
        "summary": "Set Monitor",
        "operationId": "set_monitor_api_v1_decisions__decision_id__monitor_put",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MonitorView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Notification List",
        "operationId": "notification_list_api_v1_notifications_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_NotificationPage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Read Notification",
        "operationId": "read_notification_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_NotificationView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionPlanInput": {
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expectedversion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expectedVersion"
        ],
        "title": "DecisionPlanInput"
      },
...
      "Envelope_JobView_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[JobView]"
      },
...
      "Envelope_MonitorPage_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MonitorPage"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[MonitorPage]"
      },
      "Envelope_MonitorView_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MonitorView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[MonitorView]"
      },
      "Envelope_NotificationPage_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationPage"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[NotificationPage]"
      },
      "Envelope_NotificationView_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[NotificationView]"
      },
...
      "Envelope_PlanView_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PlanView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[PlanView]"
      },
      "Envelope_PositionDecisionPage_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PositionDecisionPage"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[PositionDecisionPage]"
      },
      "Envelope_PositionDecision_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PositionDecision"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[PositionDecision]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "PARTIAL",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ],
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "cancellationRequested": {
            "type": "boolean",
            "title": "Cancellationrequested"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "stage",
          "createdAt",
          "updatedAt",
          "cancellationRequested",
          "result",
          "errorCode"
        ],
        "title": "JobView"
      },
...
      "Meta": {
        "properties": {
          "requestId": {
            "type": "string",
            "title": "Requestid"
          },
          "asOf": {
            "type": "string",
            "format": "date-time",
            "title": "Asof"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Meta"
      },
      "MonitorPage": {
        "properties": {
          "monitors": {
            "items": {
              "$ref": "#/components/schemas/MonitorView"
            },
            "type": "array",
            "title": "Monitors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "monitors"
        ],
        "title": "MonitorPage"
      },
      "MonitorUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "MonitorUpdate"
      },
...
      "PositionDecision": {
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "position-decision.v1",
            "title": "Schemaversion",
            "default": "position-decision.v1"
          },
          "decisionId": {
            "type": "string",
            "title": "Decisionid"
          },
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "instrumentId": {
            "type": "string",
            "pattern": "^(SH|SZ|BJ)\\.\\d{6}$",
            "title": "Instrumentid"
          },
          "contextId": {
            "type": "string",
            "title": "Contextid"
          },
          "asOf": {
            "type": "string",
            "format": "date-time",
            "title": "Asof"
          },
          "validUntil": {
            "type": "string",
            "format": "date-time",
            "title": "Validuntil"
          },
          "accountVersion": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Accountversion"
          },
          "releaseId": {
            "type": "string",
            "title": "Releaseid"
          },
          "assessmentIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assessmentids"
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "UNAVAILABLE"
            ],
            "title": "Status"
          },
          "action": {
            "type": "string",
            "enum": [
              "HOLD",
              "ADD",
              "REDUCE",
              "EXIT",
              "NONE"
            ],
            "title": "Action"
          },
          "reasonCodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Reasoncodes"
          },
          "evidenceIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidenceids"
          },
          "currentQuantityShares": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentquantityshares"
          },
          "targetQuantityShares": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetquantityshares"
          },
          "deltaQuantityShares": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": -1000000000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltaquantityshares"
          },
          "expectedDeltaReturnVsHold": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": -10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expecteddeltareturnvshold"
          },
          "q10DeltaReturnVsHold": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": -10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Q10Deltareturnvshold"
          },
          "q50DeltaReturnVsHold": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": -10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Q50Deltareturnvshold"
          },
          "q90DeltaReturnVsHold": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": -10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Q90Deltareturnvshold"
          },
          "stopHazard": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stophazard"
          },
          "support": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Support"
          },
          "quantTrend": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "BULLISH",
                  "NEUTRAL",
                  "BEARISH",
                  "UNCERTAIN"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Quanttrend"
          },
          "agentThesisStatus": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "SUPPORTED",
                  "WEAKENED",
                  "INVALIDATED",
                  "UNCERTAIN"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentthesisstatus"
          },
          "agentUncertainties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Agentuncertainties"
          },
          "counterEvidenceIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Counterevidenceids"
          },
          "executionPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executionpath"
          },
          "priceLower": {
            "anyOf": [
              {
                "type": "string",
                "decimal_places": 4,
                "gt": 0,
                "max_digits": 18
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricelower"
          },
          "priceUpper": {
            "anyOf": [
              {
                "type": "string",
                "decimal_places": 4,
                "gt": 0,
                "max_digits": 18
              },
              {
                "type": "null"
              }
            ],
            "title": "Priceupper"
          },
          "priceBasis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricebasis"
          },
          "triggerConditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Triggerconditions"
          },
          "estimatedCosts": {
            "anyOf": [
              {
                "type": "string",
                "decimal_places": 2,
                "max_digits": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedcosts"
          },
          "riskPolicyVersion": {
            "type": "string",
            "title": "Riskpolicyversion"
          },
          "quantityRuleVersion": {
            "type": "string",
            "title": "Quantityruleversion"
          },
          "feePolicyVersion": {
            "type": "string",
            "title": "Feepolicyversion"
          },
          "modelPredictionRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelpredictionref"
          },
          "agentContributionRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentcontributionref"
          },
          "decisionReason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Decisionreason"
          },
          "reviewAfter": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewafter"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decisionId",
          "accountId",
          "instrumentId",
          "contextId",
          "asOf",
          "validUntil",
          "accountVersion",
          "releaseId",
          "assessmentIds",
          "status",
          "action",
          "reasonCodes",
          "evidenceIds",
          "riskPolicyVersion",
          "quantityRuleVersion",
          "feePolicyVersion",
          "decisionReason"
        ],
        "title": "PositionDecision"
      },
      "PositionDecisionPage": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/PositionDecision"
            },
            "type": "array",
            "title": "Decisions"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decisions",
          "nextCursor"
        ],
        "title": "PositionDecisionPage"
      },
      "PositionEvaluationInput": {
        "properties": {
          "instrumentId": {
            "type": "string",
            "pattern": "^(SH|SZ|BJ)\\.\\d{6}$",
            "title": "Instrumentid"
          },
          "expectedVersion": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expectedversion"
          },
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrumentId",
          "expectedVersion",
          "reason"
        ],
        "title": "PositionEvaluationInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}